Refactor ScreenAsserter to new package and update references in the codebase.
